Update read_data method
Renamed `dataFile` method to `read_data`. Pretty much took out all of
the internal code and converted it into a single Pandas.read_csv call.
The new DataFrame is called simply `data`, which will simplify a lot of
things later. But this also breaks a lot of code along the way.