Initial fix to run_fit
I made some modifications to `run_fit` and `_residTotal` that made it
possible for the fitting to complete without errors.
1. Removed the addition of t=0 if it is not in the original time data.
(May want to add this back later.)
2. Changed how the temporary parameter lists are created for simulating
the ODES/fitting. I needed to do this using a bunch of Pandas code
rather than the method `_paramConvert`, which will need to be removed.
3. Changed how the residual array is calculated. It is more elegant now
using Pandas
4. Removed data normalization from the fitting method.
5. Changed how the inital fitting parameters are extracted from the
guess/fixed parameter table.
6. Add the final fit parameters to the `params` DataFrame.
7. Comment out all statistics for now.