General cleanup
I removed a whole bunch of stuff that shouldn't be important anymore, or
I will not need anymore. Some of this was helper functions that are no
longer necessary b/c of structural changes, and some functions are not
necessary b/c of Jupyter Notebooks, which weren't really a thing when I
started this project.