Rm old style str formatting, _ode_function_gen
The method _ode_function_gen was using a bunch of old style function
formatiing. I removed this formatting and refactored the code in some
places. Added a couple of extra comments.
WARNING: I realize that this is using some string replacement for cpd
names, which could cause problems later. I added a note into the code.