Add initial files for conda build (BROKEN)
I added the meta.yaml and conda_build_config.yaml files that are
necessary for running `conda build`. Unfortunately, these aren't working
right now, so I'll have to come back to this later. Something is going
on with the build process, because conda is not able to find setup.py.
Looking in the wrong directory.