Fix _rxnRate method
A lot was done here, but I updated the rxnRate function to make it a little easier to follow what is going on.  In addition, it handles stoichiometry coefficients much better, which removed a little bit of code from _halfRxn method as well. Added a _buildOde function to abstract a little stuff here.