Update read_rxns method
Did some final tweaks on the read_rxns method. The variable names wer confusing, so I updated those. I also fixed the method _ode_function_gen to fix a problem with compounds that are sub-components of another compound. In this case, I wrapped the compound names with square brackets. There is still a possiblity of errors, but much less now.